qcacmn: Scan partner links of non-Tx MBSSID for connect
Driver doesn't send unicast (UC) probe request to BSS which is
a non-Tx MBSSID profile in MBSSID set, some of the reasons for
that are:

   1) Few IoT APs may not respond at all to probe requests,
      this results in join failure after repeated probe requests.
   2) The ML-Probe response for ML-Probe request sent with non-zero
      per-STA profile will not have any per-STA profile as the
      probe response is sent by Tx-BSSID. Due to this, current
      implementation will remove all the partner links from
      connection and downgrades to SLO.

To avoid the limitations mentioned above, driver continues to
send join success without sending probe request and proceeds to
connection with existing scan entries for both assoc and partner
links.

Add implementation to prepare list of scan entries which are non-Tx
BSSID in MBSSID set whose atleast one valid partner link's scan
entry is not found. The number of entries to check from the filtered
scan list can be controlled by "num_nontx_to_scan_from_top" INI.

If the INI is set to non-zero value, then those number of entries
will be added to scan list and entries beyond this value are marked
as invalid for their partner links. Driver then prepares the scan
request with entries from scan list and continue to connect post
getting new candidate list from the scan db.

CRs-Fixed: 3925438

+++ umac/mlme/connection_mgr/core/src/wlan_cm_bss_scoring.c
@@ -3191,7 +3191,7 @@ cm_find_and_remove_dup_candidate(struct scan_cache_node *cur_scan_node,
 	uint8_t i, num_links;
 	struct partner_link_info *link_info;
 	struct scan_cache_node *tmp_scan_node;
-	uint64_t cur_entry_link_map, next_entry_link_map;
+	uint16_t cur_entry_link_map, next_entry_link_map;
 	qdf_list_node_t *cur_node = input_node, *next_node = NULL;
 
 	if (qdf_is_macaddr_zero(&cur_scan_node->entry->ml_info.mld_mac_addr))
